Refactor application structure and add new components
- Updated import paths for Translate, Branches, and NavBarTranslate components to reflect new directory structure.
- Added favicon and logo images to assets.
- Created BranchCard and BranchChart components for better separation of concerns.
- Removed old Branches component and replaced it with a new page structure using useBranches hook.
- Introduced Loader component for consistent loading state representation.
- Added NavBar styles for improved UI.
- Implemented Translate component with enhanced functionality for translation management.
- Created custom hooks for fetching branches and handling API requests.
- Added utility functions for date formatting and string checks.
