+++ src/App.js
@@ -2,10 +2,10 @@ import React from "react";
 import { Routes, Route, HashRouter } from "react-router-dom";
 import Login from "./components/Login";
 import Callback from "./components/Callback";
-import Translate from "./components/Translate";
+import Translate from "./components/Translate/Translate";
 import Changed from "./components/Changed";
-import Branches from "./components/Branches";
-import NavBarTranslate from "./components/NavBarTranslate";
+import Branches from "./pages/Branches";
+import NavBarTranslate from "./components/NavBar/NavBarTranslate";
 import "bootstrap/dist/css/bootstrap.min.css";
 import "./App.css";
