+++ src/components/NavBar/NavBarTranslate.js
