+++ src/utils/stringUtils.js
@@ -1,0 +1,1 @@
+export const isEmpty = (str) => !str || str === "to be filled in";
